Add files via upload
                                                        ****************** steps ******************
This code ENCODES your name and also DECODES your name 
All you need to do is to choose 1 for ENCODING and write your name and it will Encode it for you. 
And for DECODING choose 2  copy that encoded name and past. It will Decode for you
And for two words it just reverse it and when you decode that name it will again reverse that words.